requirements.txt
flask
langchain
langchain-core
langchain-openai
langchain-community
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
msal
requests
pytest
